+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+## master
+
+## 0.8.2
+
+* Performance improvements.
+* jest now uses `chalk` instead of its own colors implementation.
+
+## 0.8.1
+
+* `--bail` now reports with the proper error code.
+* Fixed loading of the setup file when using jasmine2.
+* Updated jsdom to 7.2.0.
+
+## 0.8.0
+
+* Added optional support for jasmine2 through the `testRunner` config option.
+* Fixed mocking support for Map, WeakMap and Set.
+* `node` was added to the defaults in `moduleFileExtensions`.
+* Updated the list of node core modules that are properly being recognized by
+  the module loader.
+
+## 0.7.1
+
+* Correctly map `process.on` into jsdom environments, fixes a bug introduced in
+  jest 0.7.0.
+
 ## 0.7.0
 
 * Fixed a memory leak with test contexts. Jest now properly cleans up